% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citeLF.R
\name{citeLF}
\alias{citeLF}
\title{Cite LANDFIRE data products}
\usage{
citeLF(lf_api, check_updates = TRUE)
}
\arguments{
\item{lf_api}{A landfire_api object}

\item{check_updates}{logical. If TRUE the the file will download the most
recent data product information. The}
}
\description{
Cite LANDFIRE data products
}
