% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landfireAPI.R
\name{.fmt_editrules}
\alias{.fmt_editrules}
\title{Internal: Format edit_rule for API call}
\usage{
.fmt_editrules(rules)
}
\arguments{
\item{rules}{A list of character vectors ordered "operator class"
"product", "operator", "value". Limited to fuel theme products only.}
}
\value{
Character vector (length = 1) with formated edit rules call
}
\description{
Internal: Format edit_rule for API call
}
\examples{
\dontrun{
edit_rule <- list(c("condition","ELEV2020","lt",500), c("change", "140CC", "st", 181))
.fmt_editrule(edit_rule)
}
}
