% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landfireAPI.R
\name{landfireAPI}
\alias{landfireAPI}
\title{Call the LANDFIRE Product Service (LFPS) API}
\usage{
landfireAPI(
  products,
  aoi,
  projection = NULL,
  resolution = NULL,
  edit_rule = NULL,
  edit_mask = NULL,
  path = NULL,
  max_time = 1000,
  method = "curl",
  verbose = TRUE
)
}
\arguments{
\item{products}{Product names as character vector
(see: https://lfps.usgs.gov/helpdocs/productstable.html)}

\item{aoi}{Area of interest as character or numeric vector defined by
latitude and longitude in decimal degrees in WGS84 and ordered
`xmin`, `ymin`, `xmax`, `ymax` or a LANDFIRE map zone.}

\item{projection}{Optional. A numeric value of the WKID for the output projection
Default is a localized Albers projection.}

\item{resolution}{Optional. A numeric value between 31-9999 specifying the
resample resolution in meters. Default is 30m.}

\item{edit_rule}{Optional. **Not currently functional**}

\item{edit_mask}{Optional. **Not currently functional**}

\item{path}{Path to `.zip` directory. Passed to `utils::download.file()`.
If NULL, a temporary directory is created.}

\item{max_time}{Maximum time, in seconds, to wait for job to be completed.}

\item{method}{Passed to `utils::download.file()`. See `?download.file`}

\item{verbose}{If FALSE suppress all status messages}
}
\value{
Returns API call passed from httr::get(). Downloads files to `path`
}
\description{
`landfireAPI` downloads LANDFIRE data by calling the LFPS API
}
\examples{
\dontrun{
products <-  c("ASP2020", "ELEV2020", "SLPP2020")
aoi <- c("-123.7835", "41.7534", "-123.6352", "41.8042")
projection <- 6414
resolution <- 90
save_file <- tempfile(fileext = ".zip")
test <- landfireAPI(products, aoi, projection, resolution, path = save_file)
}
}
