% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlandfire-package.R
\docType{package}
\name{rlandfire-package}
\alias{rlandfire}
\alias{rlandfire-package}
\title{rlandfire: Tools for Accessing and Working with LANDFIRE}
\description{
An R interface to call the LANDFIRE Product Services (LFPS) API to download LANDFIRE data products with a user specified resolution and projection.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bcknr/rlandfire}
  \item Report bugs at \url{https://github.com/bcknr/rlandfire/issues}
}

}
\keyword{internal}
