% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{viewProducts}
\alias{viewProducts}
\title{View LFPS products table}
\usage{
viewProducts()
}
\description{
\code{viewProducts()} opens the LFPS products table in your web browser
}
\examples{
viewProducts()

}
