% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAOI.R
\name{getAOI}
\alias{getAOI}
\title{Create extent vector for `landfireAPI()`}
\usage{
getAOI(data, extend = NULL)
}
\arguments{
\item{data}{A SpatRaster, SpatVector, sf, stars, or RasterLayer (raster) object}

\item{extend}{Optional. A numeric vector of 1, 2, or 4 elements passed to
[terra::extend()] by which to increase the extent by.}
}
\value{
Returns an extent vector ordered `xmin`, `ymin`, `xmax`, `ymax`
   with a lat/lon projection.
}
\description{
`getAOI` creates and extent vector in WGS84 from spatial data
}
\examples{
r <- terra::rast(nrows = 50, ncols = 50, xmin = -30, xmax = 10, ymin = -30,
  ymax = 10, crs = terra::crs("epsg:4326"), vals = rnorm(2500))
ext <- getAOI(r, c(10, 15))

}
