% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkStatus.R
\name{healthCheck}
\alias{healthCheck}
\title{Check if the LFPS API is available}
\usage{
healthCheck()
}
\value{
NULL. Prints a message to the console about the current LFPS status.
}
\description{
\code{healthCheck()} checks if the LPFS API is available
}
\examples{
\dontrun{
healthCheck()
}
}
