% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlandfire-package.R
\docType{package}
\name{rlandfire-package}
\alias{rlandfire}
\alias{rlandfire-package}
\title{rlandfire: Interface to 'LANDFIRE Product Service' API}
\description{
Provides access to a suite of geospatial data layers for wildfire management, fuel modeling, ecology, natural resource management, climate, conservation, etc., via the 'LANDFIRE' (\url{https://www.landfire.gov/index.php}) Product Service ('LFPS') API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bcknr/rlandfire}
  \item Report bugs at \url{https://github.com/bcknr/rlandfire/issues}
}

}
\author{
\strong{Maintainer}: Mark Buckner \email{mab677@cornell.edu} (\href{https://orcid.org/0000-0002-9692-7454}{ORCID}) [copyright holder]

}
\keyword{internal}
